backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0

# File handling and storage
Pillow==10.0.1
boto3==1.29.3
django-storages==1.14.2

# Async tasks
celery==5.3.4
redis==5.0.1

# Environment variables
python-decouple==3.8

# AI/NLP Libraries
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.2

# Image processing
opencv-python==4.8.1.78
pytesseract==0.3.10

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development
django-debug-toolbar==4.2.0 